chore(tests): phase 0 hygiene - single pytest config, shared conftest
tests/pytest.ini was not merely redundant, it was winning: running `pytest
tests/` set rootdir to tests/ and used that file, so pyproject.toml's coverage
addopts were silently ignored. Only CI, which passes --cov explicitly, was
measuring anything. Deleting it makes pyproject.toml the configfile.

Adds tests/conftest.py with the three fixtures that were declared identically
in two modules each (concrete_c25, steel_b500s, steel) and sets the Agg
matplotlib backend once for the suite, replacing the hand-rolled
matplotlib.use(Agg) in four modules and the import ordering it forced.

beam_example_imperial is deliberately NOT hoisted: test_beam.py and
test_rebar.py define it with the same name but different geometry and settings,
so sharing it would silently change one module's expected values. Documented in
CLAUDE.md and the roadmap.

Moves tests/modules_testing.py (829 lines, no tests) to scripts/. Nothing
imported it. build/, docs/build/ and the generated .docx/.xlsx turned out to be
untracked already, so those two roadmap items needed no change.

Suite unchanged at 733 passed, 1 skipped; ruff and mypy clean.
