+++ tests/test_rectangular.py
@@ -2,18 +2,8 @@ import pytest
 from typing import Generator
 
 from mento.material import SteelBar, Concrete
-from mento.units import inch, cm, mm, MPa
+from mento.units import inch, cm, mm
 from mento.rectangular import RectangularSection  # Assuming this is the module name
-
-
-@pytest.fixture()
-def concrete_c25() -> Concrete:
-    return Concrete("C25")
-
-
-@pytest.fixture()
-def steel_b500s() -> SteelBar:
-    return SteelBar(name="B500S", f_y=500 * MPa)
 
 
 @pytest.fixture()
